nofar's code addition, properly placed
