+++ devices_config.py
